correcao da model e do repository

+++ app/Repositories/ProductRepository.php
@@ -7,7 +7,7 @@ class ProductRepository implements ProductRepositoryInterface
 {
     protected $model;
 
-    public function __construct(User $model)
+    public function __construct(Product $model)
     {
         $this->model = $model;
     }
@@ -19,7 +19,7 @@ class ProductRepository implements ProductRepositoryInterface
 
     public function find($id)
     {
-        return $this->model->findOrFail($id);
+        return $this->model->find($id);
     }
 
     public function create(array $data)
@@ -29,15 +29,19 @@ class ProductRepository implements ProductRepositoryInterface
 
     public function update($id, array $data)
     {
-        $product = $this->model->findOrFail($id);
-        $product->update($data);
+        $product = $this->model->find($id);
+        if ($product) {
+            $product->update($data);
+        }
         return $product;
     }
 
     public function delete($id)
     {
-        $product = $this->model->findOrFail($id);
-        $product->delete();
+        $product = $this->model->find($id);
+        if ($product) {
+            $product->delete();
+        }
         return $product;
     }
 }
